refactor(machine): Contextを純粋なデータクラスに整理
MachineProcessContextからメソッド・computedプロパティ・Randomを除去し、
共有状態を保持するだけのデータクラスにした。産出物生成はstatic
MachineOutputFactoryへ分離し、各ステートはそれを呼び出す。RecipeGuid/
EffectiveRequestPowerの算出はProcessor側へ移し、加工中の電力倍率は
ProcessingStateがインラインで算出する。
