+++ moorestech_server/Assets/Scripts/Game.Block/Blocks/Machine/State/MachineOutputFactory.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 18dfb686f4b4e4b809d77410ceef2ebc
